requirements.txt
absl-py
altair
astunparse
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
flatbuffers
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib_metadata
importlib_resources
imutils
Jinja2
jsonschema
jsonschema-specifications
keras
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
numpy
oauthlib
opencv-python
opt-einsum
packaging
pandas
pillow
pkgutil_resolve_name
protobuf
pyarrow
pyasn1
pyasn1_modules
pydeck
Pygments
python-dateutil
pytz
referencing
requests
requests-oauthlib
rich
rpds-py
rsa
six
smmap
streamlit
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
toml
toolz
tornado
typing_extensions
tzdata
urllib3
watchdog
Werkzeug
wrapt
zipp